Refactoring things...
