+++ symfony/finizens_db_communication_type.sql
@@ -1,0 +1,2 @@
+INSERT INTO finizens_db.communication_type (id, name) VALUES (1, 'call');
+INSERT INTO finizens_db.communication_type (id, name) VALUES (2, 'sms');
