+++ symfony/src/Command/ParseCommand.php
