--- symfony/src/Model/BaseCommunicationModel.php
@@ -1,68 +1,0 @@
-<?php
-
-namespace App\Model;
-
-
-class BaseCommunicationModel
-{
-    protected $type;
-    protected $phone;
-    protected $date;
-
-    /**
-     * @return mixed
-     */
-    public function getType()
-    {
-        return $this->type;
-    }
-
-    /**
-     * @param string $type
-     */
-    public function setType($type)
-    {
-        $this->type = $type;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getPhone()
-    {
-        return $this->phone;
-    }
-
-    /**
-     * @param int $phone
-     */
-    public function setPhone($phone)
-    {
-        $this->phone = $phone;
-    }
-
-    /**
-     * @return \DateTime
-     */
-    public function getDate()
-    {
-        return $this->date;
-    }
-
-    /**
-     * @param \DateTime $date
-     */
-    public function setDate($date)
-    {
-        $this->date = $date;
-    }
-
-    /**
-     * @return array
-     */
-    public function __toJson(): array
-    {
-        return get_object_vars($this);
-    }
-
-}
