--- symfony/src/Model/SmsModel.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace App\Model;
-
-
-class SmsModel extends BaseCommunicationModel
-{
-
-}
